Move workstation_config_loader into chef-config
